--- src/bills2.yaml
@@ -1,27 +1,0 @@
-- month: 'May 2024'
-  readings:
-    Papa: 39399
-    Jack: 26003
-    Ian: 26460
-    Ajin: 10926
-  electric: 5734.76
-  water: 578.41
-  internet: 2000
-- month: 'April 2024'
-  readings:
-    Papa: 38910
-    Jack: 23659
-    Ian: 25042
-    Ajin: 10646
-  electric: 4992.26
-  water: 485.85
-  internet: 2000
-- month: 'March 2024'
-  readings:
-    Papa: 38372
-    Jack: 21389
-    Ian: 23576
-    Ajin: 10330
-  electric: 3654.53
-  water: 418.97
-  internet: 2000
