--- src/data.py
@@ -1,4 +1,0 @@
-
-
-data = [{"month":"May", "electric": 5000, "water": 2000, "readings": {"Papa": 1000, "Mama": 2000, "Anak": 3000}},
-        {"month":"June", "electric": 5000, "water": 2000, "readings": {"Papa": 1000, "Mama": 2000, "Anak": 3000}},]
